// src/types/Detail.ts
export interface DrugInfo {
  id: number;
  entp_name: string;
  item_name: string;
  item_seq: string;
  efcy_qesitm: string | null;
  use_method_qesitm: string | null;
  atpn_warn_qesitm: string | null;
  intrc_qesitm: string | null;
  se_qesitm: string | null;
  deposit_method_qesitm: string | null;
  item_image: string | null;
  created_at: string;
  updated_at: string;
}